Классы KeyboardEventHistory и MouseEventHistory получают события от Frame
