+++ src/main/java/tow/engine/Global.java
@@ -5,8 +5,8 @@ import tow.engine.cycle.Engine;
 import tow.engine.logger.AggregateLogger;
 import tow.engine.resources.audios.AudioStorage;
 import tow.engine.implementation.*;
-import tow.engine.input.mouse.MouseHandler;
-import tow.engine.input.keyboard.KeyboardHandler;
+import tow.engine.ui.mouse.MouseHandler;
+import tow.engine.ui.keyboard.KeyboardHandler;
 import tow.engine.map.Room;
 import tow.engine.net.client.Ping;
 import tow.engine.net.client.tcp.TCPControl;
