+++ src/main/java/tow/engine/ui/mouse/MouseEventHistory.java
@@ -1,0 +1,27 @@
+package tow.engine.ui.mouse;
+
+import org.liquidengine.legui.event.MouseClickEvent;
+import tow.engine.Global;
+
+import java.util.LinkedList;
+import java.util.List;
+
+public class MouseEventHistory {
+
+    private List<MouseClickEvent> eventHistory = new LinkedList<>();
+
+    public MouseEventHistory(){
+        //Создание обратного вызова для фиксирования всех событий мыши (кроме GUI)
+        Global.engine.render.frame.getContainer().getListenerMap().addListener(MouseClickEvent.class, event -> {
+            eventHistory.add(event);
+        });
+    }
+
+    public List<MouseClickEvent> getList(){
+        return List.copyOf(eventHistory);
+    }
+
+    public void update(){
+        eventHistory.clear();
+    }
+}
